Updated wails cli + added build tags for linux vs windows
- broke the `runElevated` method into two files for handling linux vs windows packages
